asr_tts/requirements.txt
openai-whisper
TTS
sounddevice
soundfile
